Moved a lot of row related state from Export into QueryBuilder
